added valve voltage offset settings

+++ utilities/initialize_onboard_memory/allSettings.h
@@ -52,8 +52,8 @@ class sensorSettings
 class valveSettings
 {
   public:
-    int valveOffset[2] = {220, 220};
-    int valveMax[2] = {255, 255};
+    uint8_t valveOffset[2] = {220, 220};
+    uint8_t valveMax[2] = {255, 255};
 };
